+++ packages/nextjs/src/middleware/redirects-middleware.test.ts
@@ -2206,6 +2206,41 @@ describe('RedirectsMiddleware', () => {
         expect(finalRes.status).to.equal(res.status);
       });
 
+      it('should preserve basePath if configured ', async () => {
+        const req = createRequest({
+          nextUrl: {
+            pathname: '/old-page',
+            search: '?param1=value1&param2=value2',
+            basePath: '/test',
+          },
+        });
+        const res = createResponse();
+
+        const redirectRes = createResponse({
+          redirected: true,
+          status: 302,
+          url: 'http://localhost:3000/test/new-page',
+        });
+        nextRedirectStub.returns(redirectRes);
+
+        const { middleware } = createMiddleware({
+          pattern: '/old-page',
+          target: '/new-page',
+          redirectType: REDIRECT_TYPE_301,
+          isQueryStringPreserved: true,
+        });
+
+        const finalRes = await middleware.handle(req, res);
+
+        expect(nextRedirectStub.calledOnce).to.be.true;
+        const redirectUrl = nextRedirectStub.getCall(0).args[0];
+        expect(redirectUrl).to.include('/new-page');
+        expect(redirectUrl).to.include('param1=value1');
+        expect(redirectUrl).to.include('param2=value2');
+
+        expect(finalRes).to.deep.equal(redirectRes);
+      });
+
       it('should redirect regardless of case in pattern and target', async () => {
         // Set up a clone function (used by both req and res)
         const cloneUrl = () => Object.assign({}, req.nextUrl);