+++ packages/nextjs/src/middleware/redirects-middleware.ts
@@ -223,9 +223,12 @@ export class RedirectsMiddleware extends MiddlewareBase {
             url.origin
           );
 
+          const basePath = url.basePath; // setting NextUrl.href overrides basePath, so we need to store it
           url.href = prepareNewURL.href;
           url.pathname = prepareNewURL.pathname;
           url.search = prepareNewURL.search;
+          url.basePath = basePath;
+
           if (!isAppRouterRequest) {
             url.locale = req.nextUrl.locale;
           }
@@ -389,9 +392,11 @@ export class RedirectsMiddleware extends MiddlewareBase {
 
     const newUrl = new URL(`${url.pathname.toLowerCase()}?${newQueryString}`, url.origin);
 
+    const basePath = url.basePath; // setting NextUrl.href overrides basePath, so we need to store it
     url.search = newUrl.search;
     url.pathname = newUrl.pathname.toLocaleLowerCase();
     url.href = newUrl.href;
+    url.basePath = basePath;
 
     return url;
   }